fix : change money type to double precision

+++ docs/migration/accounts.sql
@@ -6,6 +6,6 @@ CREATE TABLE IF NOT EXISTS accounts(
     client_name varchar(200) NOT NULL,
     client_firstname varchar(200) NOT NULL,
     birth_date date NOT NULL CHECK (EXTRACT(YEAR FROM AGE(current_date, birth_date)) >= 21),
-    net_monthly_salary decimal(10, 2) NOT NULL,
+    net_monthly_salary double precision NOT NULL,
     id_transaction integer REFERENCES transactions(id_transaction)
 );
